update more testing
- updated ctest controller method to be void for delete
- update corresponding ctest repository + serviceimpl methods
- removed redundant branches in emailerserviceimpl
- added new tests for restaurantintegrationtest
- update measure service impl for delete method

+++ APICode/src/main/java/com/app/APICode/measure/MeasureServiceImpl.java
@@ -63,11 +63,11 @@ public class MeasureServiceImpl implements MeasureService {
     @Override
     @Transactional
     public void deleteMeasure(String measureType) {
-        try {
+        if (measures.existsByMeasureType(measureType)) {
             measures.deleteByMeasureType(measureType);
-        } catch (EmptyResultDataAccessException e) {
-            throw new MeasureNotFoundException(measureType);
+            return;
         }
+        throw new MeasureNotFoundException(measureType);
     }
 
 }
